requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chirpstack-api==4.11.1
click==8.1.8
Flask==3.1.0
flask-sock==0.7.0
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-auth==2.38.0
googleapis-common-protos==1.69.2
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
paho-mqtt==2.1.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.3
rsa==4.9
simple-websocket==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
